# Avoid allocation to get IP bytes