Small allocation that can be avoided with a stackalloc. This optimization gets burried by the deserialization of large response object. Though, if the response is small, then it matters.

I ran the benchmark code from https://github.com/maxmind/MaxMind-DB-Reader-dotnet/pull/210 but replacing the response with an empty object.
```csharp
_reader.Find<Empty>(IPAddress.Loopback);
```

| Method | Mean     | Error   | StdDev  | Gen0   | Allocated |
|------- |---------:|--------:|--------:|-------:|----------:|
| Before   | 235.4 ns | 0.68 ns | 0.63 ns | 0.0038 |      64 B |
| After | 232.8 ns | 0.24 ns | 0.19 ns | 0.0019 |      32 B |